requirements/base.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
django-cors-headers>=4.0
django-redis>=5.4
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
drf-yasg>=1.21
django-celery-beat>=2.6
django-celery-results>=2.5
requests>=2.31
openai>=1.0.0
gunicorn>=21.0.0
